// frontend/src/api/model/administrationStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */

export type AdministrationStatus = typeof AdministrationStatus[keyof typeof AdministrationStatus];


export const AdministrationStatus = {
  GIVEN: 'GIVEN',
  OMITTED: 'OMITTED',
  VOID: 'VOID',
} as const;